Flatten package layout
Move the polythene implementation out of src/, remove the temporary loaders, and update packaging metadata for the direct package.
